@@ -69,29 +67,29 @@
     "clean-webpack-plugin": "^4.0.0",
     "cross-env": "^7.0.3",
     "css-loader": "^6.5.1",
-    "css-minimizer-webpack-plugin": "^3.1.1",
+    "css-minimizer-webpack-plugin": "^3.3.0",
     "eslint": "^7.32.0",
     "eslint-plugin-babel": "^5.3.1",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-promise": "^5.1.1",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-promise": "^5.2.0",
+    "eslint-plugin-react": "^7.27.1",
+    "eslint-plugin-react-hooks": "^4.3.0",
     "fsevents": "^2.3.2",
     "ignore-styles": "^5.0.1",
     "jsdom": "^17.0.0",
     "jsdom-global": "^3.0.2",
-    "mini-css-extract-plugin": "^2.4.4",
-    "mocha": "^8.4.0",
-    "postcss": "^8.3.11",
-    "sass": "^1.43.4",
-    "sass-loader": "^12.3.0",
+    "mini-css-extract-plugin": "^2.4.5",
+    "mocha": "^9.1.3",
+    "postcss": "^8.4.5",
+    "sass": "^1.45.0",
+    "sass-loader": "^12.4.0",
     "style-loader": "^3.3.1",
-    "terser-webpack-plugin": "^5.2.4",
+    "terser-webpack-plugin": "^5.3.0",
     "ts-node": "^10.4.0",
-    "typescript": "^4.4.4",
-    "webpack": "^5.61.0",
+    "typescript": "^4.5.4",
+    "webpack": "^5.65.0",
     "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^3.11.2"
+    "webpack-dev-server": "^3.11.3"
   },
   "browser": {
     "fs": false
